+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# rconnection 0.1.1  2024-04-26
+
+* `vfile()` can now read/write to other connections as well as files
+
 # rconnection 0.1.0  2024-04-15
 
 Initial release